+++ state_evolution/recursion_parts/full_R_recursion.py
@@ -20,7 +20,7 @@ def R_recursion(
     integral_size=5.5,
 ):
     device = S_t_tensor.device
-    dtype = S_t_tensor.dtype if torch.is_floating_point(S_t_tensor) else torch.float32
+    dtype = S_t_tensor.dtype if torch.is_floating_point(S_t_tensor) else torch.float64
 
     if R_00_sqrtm_inv_tensor is None:
         R_00_sqrt = _matrix_sqrt(R_00_tensor)
